유저 회원가입, 기업 회원가입 메서드를 유저/컴퍼니서비스에서 auth서비스로 이동
