+++ albaing-api/src/main/java/com/jobjob/albaing/service/CompanyService.java
@@ -4,8 +4,8 @@ import com.jobjob.albaing.dto.Company;
 
 
 public interface CompanyService {
-    // 기업 회원가입
-    void registerCompany(Company company);
+
+
 
 
 }
