+++ albaing-api/src/main/java/com/jobjob/albaing/service/CompanyServiceImpl.java
@@ -12,10 +12,7 @@ public class CompanyServiceImpl implements CompanyService {
     @Autowired
     private CompanyMapper companyMapper;
 
-    @Override
-    public void registerCompany(Company company) {
-        companyMapper.registerCompany(company);
-    }
+
 
 
 }
