+++ albaing-api/src/main/java/com/jobjob/albaing/service/UserService.java
@@ -5,7 +5,6 @@ import com.jobjob.albaing.dto.User;
 import java.util.Map;
 
 public interface UserService {
-    // 유저 회원가입
-    void registerUser(User user);
+
 
 }
