+++ albaing-api/src/main/java/com/jobjob/albaing/service/UserServiceImpl.java
@@ -1,22 +1,15 @@
 package com.jobjob.albaing.service;
 
-import com.jobjob.albaing.dto.User;
-import com.jobjob.albaing.mapper.UserMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 
 @Service
 public class UserServiceImpl implements UserService {
 
-    @Autowired
-    private UserMapper userMapper;
 
 
-    @Override
-    public void registerUser(User user) {
-        userMapper.registerUser(user);
-    }
+
+
 
 
 }
